refactor code, combine registerUser, registerTeacher, and registerStudent into a single controller
