--- src/controllers/teacherController.js
@@ -1,28 +1,0 @@
-const User = require("../models/User.js");
-
-const registerTeacher = async (req, res) => {
-  try {
-    const { firstName, lastName, email, password } = req.body;     
-
-    const existingUser = await User.findOne({ email });
-    if (existingUser) {
-      return res.status(400).json({ message: "User already exists" });
-    }
-    const newTeacher = new User({
-      firstName,
-      lastName,
-      email,
-      password, // Directly use the password from req.body
-      role: 'teacher'
-    });
-    
-    await newTeacher.save();
-
-    res.status(201).json({ message: "Teacher registered successfully" });
-  } catch (error) {
-    console.error("Error registering teacher:", error);
-    res.status(500).json({ message: "Internal server error", error: error.message });
-  }
-};
-
-module.exports = { registerTeacher };
