+++ src/routes/userRouter.js
@@ -1,17 +1,12 @@
-const express = require("express");
+const express = require('express');
 const router = express.Router();
-const userController = require('../controllers/userController');
-const studentController = require('../controllers/studentController');
-const teacherController = require('../controllers/teacherController');
-const { validateStudent, validateTeacher } = require('../middleware/validationMiddleware');
-
-// POST /api/v1/register
-router.post("/register", userController.registerUser);
+const { registerStudent, registerTeacher } = require('../controllers/authController');
+const { validateStudent, validateTeacher } = require('../middleware/userValidation');
 
 // POST /api/v1/register/student
-router.post("/register/student", validateStudent, studentController.registerStudent);
+router.post("/register/student", validateStudent, registerStudent);
 
 // POST /api/v1/register/teacher
-router.post("/register/teacher", validateTeacher, teacherController.registerTeacher);
+router.post("/register/teacher", validateTeacher, registerTeacher);
 
 module.exports = router;
